Renamed creators to FileBuilder
